refactored pipette flows to match other flows
