+++ app/src/organisms/ModuleWizardFlows/useFilteredModuleCalibrationSteps.ts
@@ -3,6 +3,11 @@ import { useFilterWizardStepsFrom } from '../../resources/wizards/hooks'
 import type { Subsystem } from '@opentrons/api-client'
 import type { ModuleCalibrationWizardStep } from './types'
 
+export const useFilteredModuleCalibrationSteps = (
+  subsystem: Subsystem
+): ModuleCalibrationWizardStep[] =>
+  useFilterWizardStepsFrom(getModuleCalibrationSteps(), subsystem)
+
 export const getModuleCalibrationSteps = (): ModuleCalibrationWizardStep[] => {
   return [
     { section: SECTIONS.BEFORE_BEGINNING },
@@ -14,8 +19,3 @@ export const getModuleCalibrationSteps = (): ModuleCalibrationWizardStep[] => {
     { section: SECTIONS.SUCCESS },
   ]
 }
-
-export const useFilteredModuleCalibrationSteps = (
-  subsystem: Subsystem
-): ModuleCalibrationWizardStep[] =>
-  useFilterWizardStepsFrom(getModuleCalibrationSteps(), subsystem)
